fixed some minor bugs, mostly due to conversion between differnt integer types: - scan modes were not set properly - longer integration times didn't work

+++ src/labview_fpga_lib/galvo_scan/galvo_scan_wrapper.h
@@ -47,15 +47,15 @@ void set_dVmin_y(int16_t value, NiFpga_Session* session, NiFpga_Status* status);
 void set_scanmode_x(uint8_t value, NiFpga_Session* session, NiFpga_Status* status);
 void set_scanmode_y(uint8_t state, NiFpga_Session* session, NiFpga_Status* status);
 void set_detector_mode(uint8_t state, NiFpga_Session* session, NiFpga_Status* status);
-void set_settle_time(uint8_t value, NiFpga_Session* session, NiFpga_Status* status);
-uint8_t read_settle_time(NiFpga_Session* session, NiFpga_Status* status);
+void set_settle_time(uint16_t value, NiFpga_Session* session, NiFpga_Status* status);
+uint16_t read_settle_time(NiFpga_Session* session, NiFpga_Status* status);
 void set_meas_per_pt(uint8_t value, NiFpga_Session* session, NiFpga_Status* status);
 uint8_t read_meas_per_pt(NiFpga_Session* session, NiFpga_Status* status);
 
 
 // read parameters
 int32_t read_i(NiFpga_Session* session, NiFpga_Status* status);
-int16_t read_elements_written_to_dma(NiFpga_Session* session, NiFpga_Status* status);
+int32_t read_elements_written_to_dma(NiFpga_Session* session, NiFpga_Status* status);
 int16_t read_detector_signal(NiFpga_Session* session, NiFpga_Status* status);
 int32_t read_ix(NiFpga_Session* session, NiFpga_Status* status);
 int32_t read_iy(NiFpga_Session* session, NiFpga_Status* status);
